nvim: Manage plugins with submodules
lazy.nvim's lock file handling is very annoying and there's no way to
have it leave the lock file alone unless an update is explicitly
requested.

If plugin A is out-of-date on disk and plugin B is missing on disk, then
installing plugin B causes A's commit hash to change to the checked out
commit in the lock file. This makes it impossible to restore to the
desired commits without first undoing the changes to the lock file.
Additionally, if auto-install is disabled, the restore action deletes
plugin B from the lock file.

Upstream bug reports indicate that this is intentional behavior (1279,
1514, 1740), so the behavior is unlikely to ever change. Since I don't
use any of lazy.nvim's fancy functionality, just get rid of the plugin
and use submodules.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>